Add login/logout screens
ub-12

When there is no user logged in, a Login button is visible on the home page. If someone tries to create a post or perform other actions when not logged in, they will be redirected to the Login screen. The login screen presents a field for the email and password. Users need to have previously created a user account. Once logged in, the home screen will no longer show a Login screen, and instead a Logout button is visible in top right corner. Logout redirects to the login screen.