Add Create-User page
ub 14

When no user is logged in, show a login and a create-user button
on the homepage.
When a user is logged in, show a create-post button on the homepage.
Fixed other UI issues: now create-user form doesn't show the password
in plaintext but rather with asterisks, and now the user index page
doesn't show users' passwords.